backend/requirements.txt
#common
dagshub
mlflow
python-dotenv
pydantic

#backend-developer
pandas
numpy
scikit-learn
pymongo
python-box
pyyaml
# xgboost
uv
psutil
bcrypt
pyjwt
certifi
python-multipart
sendgrid

#frontend-customer

# Add testing dependencies
pytest
pytest-cov
pytest-mock
httpx  # For TestClient

